+++ packages/shared/src/types/json.ts
@@ -25,6 +25,7 @@ import type { OrganizationCustomRoleKey, OrganizationPermissionKey } from './org
 import type { OrganizationSettingsJSON } from './organizationSettings';
 import type { OrganizationSuggestionStatus } from './organizationSuggestion';
 import type { PhoneCodeChannel } from './phoneCodeChannel';
+import type { ProtectConfigJSON } from './protectConfig';
 import type { SamlIdpSlug } from './saml';
 import type { SessionStatus, SessionTask } from './session';
 import type { SessionVerificationLevel, SessionVerificationStatus } from './sessionVerification';
@@ -90,6 +91,7 @@ export interface EnvironmentJSON extends ClerkResourceJSON {
   maintenance_mode: boolean;
   organization_settings: OrganizationSettingsJSON;
   user_settings: UserSettingsJSON;
+  protect_config: ProtectConfigJSON;
 }
 
 export type LastAuthenticationStrategy =
@@ -113,7 +115,6 @@ export interface ClientJSON extends ClerkResourceJSON {
   captcha_bypass?: boolean; // this is used by the @clerk/testing package
   last_active_session_id: string | null;
   last_authentication_strategy: LastAuthenticationStrategy | null;
-  client_trust_state?: ClientTrustState;
   cookie_expires_at: number | null;
   created_at: number;
   updated_at: number;
@@ -624,9 +625,9 @@ export interface FeatureJSON extends ClerkResourceJSON {
   object: 'feature';
   id: string;
   name: string;
-  description: string;
+  description: string | null;
   slug: string;
-  avatar_url: string;
+  avatar_url: string | null;
 }
 
 /**
@@ -637,25 +638,17 @@ export interface BillingPlanJSON extends ClerkResourceJSON {
   id: string;
   name: string;
   fee: BillingMoneyAmountJSON;
-  annual_fee: BillingMoneyAmountJSON;
-  annual_monthly_fee: BillingMoneyAmountJSON;
-  amount: number;
-  amount_formatted: string;
-  annual_amount: number;
-  annual_amount_formatted: string;
-  annual_monthly_amount: number;
-  annual_monthly_amount_formatted: string;
-  currency_symbol: string;
-  currency: string;
-  description: string;
+  annual_fee: BillingMoneyAmountJSON | null;
+  annual_monthly_fee: BillingMoneyAmountJSON | null;
+  description: string | null;
   is_default: boolean;
   is_recurring: boolean;
   has_base_fee: boolean;
   for_payer_type: BillingPayerResourceType;
   publicly_visible: boolean;
   slug: string;
-  avatar_url: string;
-  features: FeatureJSON[];
+  avatar_url: string | null;
+  features?: FeatureJSON[];
   free_trial_days?: number | null;
   free_trial_enabled?: boolean;
 }
@@ -666,13 +659,17 @@ export interface BillingPlanJSON extends ClerkResourceJSON {
 export interface BillingPaymentMethodJSON extends ClerkResourceJSON {
   object: 'commerce_payment_method';
   id: string;
-  last4: string;
-  payment_type: 'card' | 'link';
-  card_type: string;
-  is_default: boolean;
-  is_removable: boolean;
+  last4: string | null;
+  payment_type?: 'card';
+  card_type: string | null;
+  is_default?: boolean;
+  is_removable?: boolean;
   status: BillingPaymentMethodStatus;
-  wallet_type: string | null;
+  wallet_type?: string | null;
+  expiry_year?: number | null;
+  expiry_month?: number | null;
+  created_at?: number | null;
+  updated_at?: number | null;
 }
 
 /**
@@ -713,11 +710,10 @@ export interface BillingPaymentJSON extends ClerkResourceJSON {
   object: 'commerce_payment';
   id: string;
   amount: BillingMoneyAmountJSON;
-  paid_at?: number;
-  failed_at?: number;
+  paid_at: number | null;
+  failed_at: number | null;
   updated_at: number;
-  payment_method: BillingPaymentMethodJSON;
-  subscription: BillingSubscriptionItemJSON;
+  payment_method?: BillingPaymentMethodJSON | null;
   subscription_item: BillingSubscriptionItemJSON;
   charge_type: BillingPaymentChargeType;
   status: BillingPaymentStatus;
@@ -733,7 +729,6 @@ export interface BillingSubscriptionItemJSON extends ClerkResourceJSON {
   credit?: {
     amount: BillingMoneyAmountJSON;
   };
-  payment_method_id: string;
   plan: BillingPlanJSON;
   plan_period: BillingSubscriptionPlanPeriod;
   status: BillingSubscriptionStatus;
@@ -745,8 +740,7 @@ export interface BillingSubscriptionItemJSON extends ClerkResourceJSON {
   period_end: number | null;
   canceled_at: number | null;
   past_due_at: number | null;
-  // TODO(@COMMERCE): Remove optional after GA.
-  is_free_trial?: boolean;
+  is_free_trial: boolean;
 }
 
 /**
@@ -771,7 +765,7 @@ export interface BillingSubscriptionJSON extends ClerkResourceJSON {
   updated_at: number | null;
   past_due_at: number | null;
   subscription_items: BillingSubscriptionItemJSON[] | null;
-  eligible_for_free_trial?: boolean;
+  eligible_for_free_trial: boolean;
 }
 
 /**
@@ -792,15 +786,19 @@ export interface BillingCheckoutTotalsJSON {
   subtotal: BillingMoneyAmountJSON;
   tax_total: BillingMoneyAmountJSON;
   total_due_now: BillingMoneyAmountJSON;
-  credit: BillingMoneyAmountJSON;
-  past_due: BillingMoneyAmountJSON;
+  credit: BillingMoneyAmountJSON | null;
+  past_due: BillingMoneyAmountJSON | null;
+  total_due_after_free_trial: BillingMoneyAmountJSON | null;
 }
 
 /**
  * @experimental This is an experimental API for the Billing feature that is available under a public beta, and the API is subject to change. It is advised to [pin](https://clerk.com/docs/pinning) the SDK version and the clerk-js version to avoid breaking changes.
  */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface BillingStatementTotalsJSON extends Omit<BillingCheckoutTotalsJSON, 'total_due_now'> {}
+export interface BillingStatementTotalsJSON {
+  grand_total: BillingMoneyAmountJSON;
+  subtotal: BillingMoneyAmountJSON;
+  tax_total: BillingMoneyAmountJSON;
+}
 
 /**
  * @experimental This is an experimental API for the Billing feature that is available under a public beta, and the API is subject to change. It is advised to [pin](https://clerk.com/docs/pinning) the SDK version and the clerk-js version to avoid breaking changes.
@@ -817,8 +815,7 @@ export interface BillingCheckoutJSON extends ClerkResourceJSON {
   status: 'needs_confirmation' | 'completed';
   totals: BillingCheckoutTotalsJSON;
   is_immediate_plan_change: boolean;
-  // TODO(@COMMERCE): Remove optional after GA.
-  free_trial_ends_at: number | null;
+  free_trial_ends_at?: number;
   payer: BillingPayerJSON;
   needs_payment_method: boolean;
 }
@@ -829,19 +826,19 @@ export interface BillingCheckoutJSON extends ClerkResourceJSON {
 export interface BillingPayerJSON extends ClerkResourceJSON {
   object: 'commerce_payer';
   id: string;
-  created_at: number;
-  updated_at: number;
-  image_url: string | null;
+  created_at?: number;
+  updated_at?: number;
+  image_url?: string;
 
   // User attributes
-  user_id?: string;
-  email?: string;
-  first_name?: string;
-  last_name?: string;
+  user_id: string | null;
+  email?: string | null;
+  first_name?: string | null;
+  last_name?: string | null;
 
   // Organization attributes
-  organization_id?: string;
-  organization_name?: string;
+  organization_id: string | null;
+  organization_name?: string | null;
 }
 
 export interface ApiKeyJSON extends ClerkResourceJSON {
@@ -857,6 +854,10 @@ export interface ApiKeyJSON extends ClerkResourceJSON {
   expiration: number | null;
   created_by: string | null;
   description: string | null;
+  /**
+   * This property is only present in the response from `create()`.
+   */
+  secret?: string;
   last_used_at: number | null;
   created_at: number;
   updated_at: number;